изменено:      data/shaders/SSR.glsl 	изменено:      data/shaders/base_frag.glsl 	изменено:      data/shaders/blur.glsl 	изменено:      data/shaders/bng.glsl 	изменено:      data/shaders/deferred.glsl 	новый файл:    data/shaders/draw_texture_frag.glsl 	новый файл:    data/shaders/draw_texture_vert.glsl 	изменено:      data/shaders/fragment.glsl 	изменено:      data/shaders/lighting.glsl 	изменено:      data/shaders/median.glsl 	изменено:      data/shaders/predenoiser.glsl 	изменено:      data/shaders/screen_plane.glsl 	изменено:      data/shaders/sky_generator.glsl 	изменено:      data/shaders/ssao.glsl 	новый файл:    data/shaders/txaa.glsl 	новый файл:    data/shaders/txaa_ss.glsl 	изменено:      data/shaders/vertex.glsl 	изменено:      inc/structures/components/camera_component.h 	изменено:      inc/structures/components/light_component.h 	изменено:      inc/structures/framebuffer.h 	изменено:      inc/structures/list.h 	изменено:      inc/structures/scene.h 	изменено:      inc/structures/shader.h 	изменено:      inc/structures/texture.h 	изменено:      inc/structures/types.h 	изменено:      src/gl_tester_2.c 	изменено:      src/linalg.c 	изменено:      src/memmanager.c 	изменено:      src/structures/components/camera_component.c 	изменено:      src/structures/components/light_component.c 	изменено:      src/structures/framebuffer.c 	изменено:      src/structures/game_object.c 	изменено:      src/structures/material.c 	изменено:      src/structures/mesh.c 	изменено:      src/structures/scene.c 	изменено:      src/structures/shader.c 	изменено:      src/structures/texture.c

+++ src/memmanager.c
@@ -74,7 +74,7 @@ void sFree2(void** ptr)
 	{
 		return;
 	}
-    for (int i=0; i<sSizeof(ptr)/sizeof(uintptr_t); i++)
+    for (size_t i=0; i<sSizeof(ptr)/sizeof(uintptr_t); i++)
     {
         sFree(ptr[i]);
     }
